Add unittest for daemon

+++ zippy/client/daemon.py
@@ -11,18 +11,18 @@ import sys
 import tempfile
 import time
 
-import yaml
-
 from imapclient import IMAPClient
 from imapclient.exceptions import LoginError
 
 if __package__:
     # sphinx imports this module, so need to use relative imports for the purpose
     from .log import get_logger
+    from .config import get_config
 else:
     # running this as script doesnot work with relative imports as there is no parent
     # TODO: split these and put it somewhere else to remove this kind of checks
     from log import get_logger  # type: ignore  #pylint: disable=no-name-in-module
+    from config import get_config  # type: ignore  #pylint: disable=no-name-in-module
 
 
 LOGGER = get_logger()
@@ -90,7 +90,8 @@ class Daemon:
 
     def _remove_pidfile(self):
         LOGGER.info("Removing PID file...")
-        self.PID_FILE.unlink()
+        if self.PID_FILE.exists():
+            self.PID_FILE.unlink()
 
     def start(self):
         """Start daemon."""
@@ -128,6 +129,9 @@ class Daemon:
     def stop(self):
         """Kill the running daemon."""
         # Try killing the daemon process
+        if self.disable_daemon:
+            return
+
         pid = self._get_pid()
 
         if not pid:
@@ -143,13 +147,53 @@ class Daemon:
                 os.kill(pid, signal.SIGTERM)
                 time.sleep(0.2)
             except OSError:
-                if self.PID_FILE.exists():
-                    self.PID_FILE.unlink()
-                    LOGGER.info("Removed PID file")
+                self._remove_pidfile()
                 break
 
 
-def download_new_email(username, password, host, imap_port=993, timeout=3):
+def block_till_new_mail(server, timeout=60):
+    """Return true if there's new mail, else checks for timeout."""
+    failure = 0
+
+    while time.time() < timeout:
+        if failure >= 10:
+            failure = 1
+
+        try:
+            # Wait for up to 30 seconds for an IDLE response
+            responses = server.idle_check(timeout=30)
+            # if server was up, response should have come as timeout is 30sec.
+            # so, it's not required to check for socket.error here.
+        # log the unexpected exception
+        except socket.error:
+            failure += 1
+            LOGGER.info("Sleeping for %s second", failure * 10)
+            time.sleep(failure * 10)
+            continue
+        except Exception:  # pylint: disable=broad-except
+            LOGGER.exception("Unexpected error occurred.")
+            raise
+        else:
+            failure = 0
+            if responses:
+                return True
+        time.sleep(10)
+
+    try:
+        server.idle_done()
+    # nothing expected exception here. but let's log these if it occurs.
+    except Exception:  # pylint: disable=broad-except
+        LOGGER.exception("Unexpected error occurred.")
+        raise
+    else:
+        LOGGER.info("\nIDLE mode done")
+        time.sleep(10)
+        return False
+
+
+def download_new_email(
+    username, password, host, imap_port=993, timeout=3
+):  # pylint:disable=too-many-branches,too-many-statements
     """Check for new emails and download them.
 
     Parameters
@@ -179,74 +223,68 @@ def download_new_email(username, password, host, imap_port=993, timeout=3):
 
     LOGGER.debug(username)
 
-    try:
-        server = IMAPClient(
-            host, port=imap_port, ssl=True, ssl_context=ssl_context, timeout=timeout
-        )
-    except socket.error:
-        LOGGER.exception("Could not connect to the mail server.")
-        return
-    except ssl.SSLError:
-        LOGGER.exception("Error due to ssl.")
-        return
-
-    try:
-        server.login(username, password)
-        server.select_folder("INBOX")
-    except LoginError:
-        LOGGER.exception("Credentials could not be verified for '%s'.", username)
-        server.shutdown()
-        return
+    failure = 0
 
     while True:
-        # reconnect every 7 mins, as per recommendation to do so.
-        timeout = time.time() + 7 * 60  # after 7 mins
+        if failure > 10:
+            failure = 1
 
-        # Start IDLE mode
         try:
-            server.idle()
-        except Exception:  # pylint: disable=broad-except
-            LOGGER.exception("Unexpected exception occurred.")
-            return
-        else:
-            LOGGER.info(
-                "Connection is now in IDLE mode, send yourself an email or quit with ^c"
+            server = IMAPClient(
+                host, port=imap_port, ssl=True, ssl_context=ssl_context, timeout=timeout
             )
+        except socket.error:
+            LOGGER.exception("Could not connect to the mail server.")
+            failure += 1
+            LOGGER.info("Sleeping for %s second", failure * 10)
+            time.sleep(10 * failure)
+            continue
+        except ssl.SSLError:
+            LOGGER.exception("Error due to ssl.")
+            break
+        else:
+            failure = 0
 
-        curr_time = time.time()
-        while curr_time < timeout:
+        try:
+            server.login(username, password)
+            server.select_folder("INBOX")
+        except LoginError:
+            LOGGER.exception("Credentials could not be verified for '%s'.", username)
+            server.shutdown()
+            break
+        except socket.error:
+            LOGGER.exception("Could not connect to the mail server.")
+            failure += 1
+            time.sleep(10 * failure)
+            continue
+        else:
+            failure = 0
+
+        while True:
+            # reconnect every 7 mins, as per recommendation to do so.
+            timeout = int(time.time()) + 7 * 60  # after 7 mins
+
             try:
-                # Wait for up to 30 seconds for an IDLE response
-                responses = server.idle_check(timeout=30)
-                # if server was up, response should have come as timeout is 30sec.
-                # so, it's not required to check for socket.error here.
-            # log the unexpected exception
+                server.idle()
+            except socket.error:
+                LOGGER.exception("Could not connect to the mail server.")
+                break
             except Exception:  # pylint: disable=broad-except
-                LOGGER.exception("Unexpected error occurred.")
+                LOGGER.exception("Unexpected exception occurred.")
                 return
             else:
+                failure = 0
                 LOGGER.info(
-                    "Server sent: %s for %s",
-                    responses if responses else "nothing",
-                    username,
+                    "Connection is now in IDLE mode, send yourself an email or quit with ^c"
                 )
-            time.sleep(10)
 
-        try:
-            server.idle_done()
-        # nothing expected exception here. but let's log these if it occurs.
-        except Exception:  # pylint: disable=broad-except
-            LOGGER.exception("Unexpected error occurred.")
-            return
-        else:
-            LOGGER.info("\nIDLE mode done")
-            time.sleep(10)
+            has_new_mail = block_till_new_mail(server, timeout=timeout)
+            if has_new_mail:
+                pass
 
 
 if __name__ == "__main__":
-    STREAM = open(pathlib.Path(__file__).parents[2] / ".env.yml", "r")
-    CONFIG = yaml.safe_load(STREAM)
-    CLIENT_CONFIG = CONFIG.get("CLIENT")
+    CLIENT_CONFIG = get_config("CLIENT")
     if not CLIENT_CONFIG:
         print("Config not setup properly. Check env.example.yml for more info.")
         sys.exit(1)
